+++ cpp/metrics_utils.cpp
@@ -25,7 +25,7 @@ namespace metrics_utils{
         return (centered.adjoint() * centered) / (static_cast<double>(M.rows()) - 1.0);
     }
 
-    Eigen::MatrixX3d points_idx_to_matrix(cilantro::VectorSet3d const & points, std::vector<unsigned long> const & idx){
+    Eigen::MatrixX3d points_idx_to_matrix(cilantro::VectorSet3d const & points, std::vector<int> const & idx){
         Eigen::MatrixX3d matrix(idx.size(), 3);
         Eigen::Index row_idx = 0;
         for (const auto & i : idx){
@@ -58,12 +58,12 @@ namespace metrics_utils{
         return PointCloud{pc_map_points};
     }
 
-    std::vector<unsigned long> get_radius_search_indices(KDTree const & tree,
+    std::vector<int> get_radius_search_indices(KDTree const & tree,
                                                        Eigen::Vector3d const & query, double radius){
         cilantro::NeighborSet<double> nn = tree.radiusSearch(query, radius);
-        std::vector<unsigned long> indices(nn.size());
+        std::vector<int> indices(nn.size());
         for (size_t i = 0; i < nn.size(); ++i){
-            indices[i] = nn[i].index;
+            indices[i] = static_cast<int>(nn[i].index);
         }
 
         return indices;
@@ -71,7 +71,7 @@ namespace metrics_utils{
 
     namespace metrics_algorithm{
         std::optional<double> compute_eigenvalues(cilantro::VectorSet3d const & points,
-                                   std::vector<unsigned long> const & indices){
+                                   std::vector<int> const & indices){
             Eigen::MatrixX3d tmp = points_idx_to_matrix(points, indices);
             Eigen::MatrixX3d cov_matrix = cov(tmp);
             Eigen::VectorXd eigenvalues = cov_matrix.eigenvalues().real();
@@ -79,7 +79,7 @@ namespace metrics_utils{
         }
 
         std::optional<double> compute_entropy(cilantro::VectorSet3d const & points,
-                                              std::vector<unsigned long> const & indices){
+                                              std::vector<int> const & indices){
             Eigen::MatrixXd tmp = points_idx_to_matrix(points, indices);
             Eigen::MatrixXd cov_matrix = cov(tmp);
             double det =  (2. * M_PI * M_E * cov_matrix).determinant();
